+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/service/datapoint/EventDetectorPropertiesServiceTest.java
@@ -18,7 +18,6 @@ import org.scadalts.e2e.service.impl.services.pointvalue.PointValueParams;
 import org.scadalts.e2e.test.impl.creators.DataPointObjectsCreator;
 import org.scadalts.e2e.test.impl.creators.DataSourcePointObjectsCreator;
 import org.scadalts.e2e.test.impl.creators.EventDetectorObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestParameterizedWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.DataPointPropertiesAdapter;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 import org.scadalts.e2e.test.impl.utils.TestWithoutPageUtil;
@@ -32,7 +31,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 @Log4j2
-@RunWith(TestParameterizedWithPageRunner.class)
+@RunWith(Parameterized.class)
 public class EventDetectorPropertiesServiceTest {
 
     private static DataSourceCriteria DATA_SOURCE_CRITERIA = DataSourceCriteria.virtualDataSourceSecond();
@@ -79,7 +78,7 @@ public class EventDetectorPropertiesServiceTest {
     public static void setupForAll() {
         DataSourcePointCriteria dataSourcePointCriteria = DataSourcePointCriteria.criteria(DATA_SOURCE_CRITERIA,
                 DATA_POINT_CRITERIA);
-        navigationPage = TestWithPageUtil.getNavigationPage();
+        navigationPage = TestWithPageUtil.openNavigationPage();
         dataSourcePointObjectsCreator = new DataSourcePointObjectsCreator(navigationPage, dataSourcePointCriteria);
         dataSourcePointObjectsCreator.createObjects();
         dataPointCriteria = dataSourcePointCriteria.getDataPoint();
@@ -87,7 +86,8 @@ public class EventDetectorPropertiesServiceTest {
 
     @AfterClass
     public static void cleanForAll() {
-        dataSourcePointObjectsCreator.deleteObjects();
+        if(dataSourcePointObjectsCreator != null)
+            dataSourcePointObjectsCreator.deleteObjects();
     }
 
     @Before
