+++ scada-lts-e2e-test/scada-lts-e2e-test-impl/src/main/java/org/scadalts/e2e/test/impl/tests/service/storungs/live/GetActiveLivesOnChangeServiceTest.java
@@ -4,7 +4,6 @@ import lombok.extern.log4j.Log4j2;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.scadalts.e2e.page.impl.criterias.DataPointCriteria;
 import org.scadalts.e2e.page.impl.criterias.DataSourceCriteria;
 import org.scadalts.e2e.page.impl.criterias.IdentifierObjectFactory;
@@ -16,7 +15,6 @@ import org.scadalts.e2e.page.impl.pages.navigation.NavigationPage;
 import org.scadalts.e2e.service.impl.services.storungs.PaginationParams;
 import org.scadalts.e2e.service.impl.services.storungs.StorungAlarmResponse;
 import org.scadalts.e2e.test.impl.creators.StorungsAndAlarmsObjectsCreator;
-import org.scadalts.e2e.test.impl.runners.TestWithPageRunner;
 import org.scadalts.e2e.test.impl.utils.RegexUtil;
 import org.scadalts.e2e.test.impl.utils.TestWithPageUtil;
 
@@ -31,7 +29,6 @@ import static org.junit.Assert.assertThat;
 import static org.scadalts.e2e.test.impl.utils.StorungsAndAlarmsUtil.*;
 
 @Log4j2
-@RunWith(TestWithPageRunner.class)
 public class GetActiveLivesOnChangeServiceTest {
 
     private static DataPointIdentifier alarmIdentifier = IdentifierObjectFactory.dataPointAlarmBinaryTypeName();
@@ -48,7 +45,7 @@ public class GetActiveLivesOnChangeServiceTest {
         DataPointCriteria pointStorung = DataPointCriteria.noChange(storungIdentifier, "0",
                 DataPointLoggingProperties.logging(LoggingType.ON_CHANGE));
 
-        NavigationPage navigationPage = TestWithPageUtil.getNavigationPage();
+        NavigationPage navigationPage = TestWithPageUtil.openNavigationPage();
 
         storungsAndAlarmsObjectsCreator =
                 new StorungsAndAlarmsObjectsCreator(navigationPage, dataSourceCriteria, pointAlarm, pointStorung);
@@ -81,7 +78,8 @@ public class GetActiveLivesOnChangeServiceTest {
 
     @AfterClass
     public static void clean() {
-        storungsAndAlarmsObjectsCreator.deleteObjects();
+        if(storungsAndAlarmsObjectsCreator != null)
+            storungsAndAlarmsObjectsCreator.deleteObjects();
     }
 
     @Test
